+++ frontend/components/UI/index.tsx
@@ -9,3 +9,4 @@ export * from '@/components/UI/RadioGroup';
 export * from '@/components/UI/SliderDouble';
 export * from '@/components/UI/TableRow';
 export * from '@/components/UI/TabPanel';
+export * from '@/components/UI/Toggle';
