+++ frontend/components/index.tsx
@@ -1,2 +1,3 @@
 export * from '@/components/UI';
+export * from '@/components/ClearFavoritesBtn';
 export * from '@/components/Navigation';
